+++ packages/patternfly-4/react-core/src/components/Wizard/WizardBody.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+import styles from '@patternfly/patternfly/components/Wizard/wizard.css';
+import { css } from '@patternfly/react-styles';
+
+interface WizardBodyProps {
+  children?: any;
+  hasBodyPadding: boolean;
+}
+
+const WizardBody: React.SFC<WizardBodyProps> = ({ children, hasBodyPadding }) => (
+  <main className={css(styles.wizardMain, !hasBodyPadding && 'pf-m-no-padding')}>
+    {children}
+  </main>
+);
+
+export default WizardBody;