+++ packages/patternfly-4/react-core/src/components/index.ts
@@ -36,3 +36,4 @@ export * from './TextArea';
 export * from './TextInput';
 export * from './Title';
 export * from './Tooltip';
+export * from './Wizard';