+++ packages/patternfly-4/react-tokens/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-tokens@2.1.0...@patternfly/react-tokens@2.1.1) (2019-03-15)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
+# 2.1.0 (2019-03-15)
+
+
+### Features
+
+* **Table:** pass rowData and extraData to table callbacks ([#1542](https://github.com/patternfly/patternfly-react/issues/1542)) ([cb3e010](https://github.com/patternfly/patternfly-react/commit/cb3e010))
+
+
+
+
+
 ## 2.0.7 (2019-03-11)